requirements.txt
Flask==2.2.2
pyngrok==5.2.1
GitPython==3.1.30
pylint==2.16.1
requests==2.28.2
python-dotenv==0.21.1
